bugs fixed, validation

+++ server/src/main/java/bookshow/controller/UsedPropController.java
@@ -3,6 +3,7 @@ package bookshow.controller;
 import bookshow.domain.Bid;
 import bookshow.domain.props.UsedProp;
 import bookshow.domain.props.UsedPropStatus;
+import bookshow.domain.users.Role;
 import bookshow.domain.users.User;
 import bookshow.service.UsedPropService;
 import bookshow.service.UserService;
@@ -10,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
 
@@ -28,6 +30,7 @@ public class UsedPropController {
     @Autowired
     private UserService userService;
 
+    @PreAuthorize("hasAuthority('ADMINFAN')")
     @RequestMapping(
             value = "/all",
             method = RequestMethod.GET,
@@ -46,6 +49,7 @@ public class UsedPropController {
         List<UsedProp> usedProps = usedPropService.findByFanAdminIsNotNull();
         return new ResponseEntity<>(usedProps, HttpStatus.OK);
     }
+
     //all not expired and approved, and not finished
     @RequestMapping(
             method = RequestMethod.GET,
@@ -54,12 +58,13 @@ public class UsedPropController {
         List<UsedProp> usedProps = usedPropService.findByActiveUntilGreaterThanAndStatusEqualsAndAcceptedBidNull(new java.util.Date(), UsedPropStatus.APPROVED);
         return new ResponseEntity<>(usedProps, HttpStatus.OK);
     }
+
     //all exept finished
     @RequestMapping(
             value = "/not-finished",
             method = RequestMethod.GET,
             produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<List<UsedProp>> getExceptFinished(){
+    public ResponseEntity<List<UsedProp>> getExceptFinished() {
         List<UsedProp> usedProps = usedPropService.findByActiveUntilGreaterThanAndAcceptedBidNullAndStatusNot(new java.util.Date(), UsedPropStatus.DECLINED);
         return new ResponseEntity<>(usedProps, HttpStatus.OK);
     }
@@ -72,51 +77,58 @@ public class UsedPropController {
         UsedProp usedProp = usedPropService.findOne(id);
         return new ResponseEntity<>(usedProp, HttpStatus.OK);
     }
+
     //used props by user
     @RequestMapping(
             value = "/user",
             method = RequestMethod.GET,
             produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<List<UsedProp>> getUsedPropsByUser(Principal principal){
+    public ResponseEntity<List<UsedProp>> getUsedPropsByUser(Principal principal) {
         List<UsedProp> usedProps = usedPropService.findByUsername(principal.getName());
-        return new ResponseEntity<>(usedProps,HttpStatus.OK);
+        return new ResponseEntity<>(usedProps, HttpStatus.OK);
     }
+
+    @PreAuthorize("hasAuthority('ADMINFAN')")
     @RequestMapping(
-            value = "/accept-decline/{id}",
+            value = "/approve-decline/{id}",
             method = RequestMethod.GET,
             produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<UsedProp> getRefuseAccept(@PathVariable Long id, @RequestParam("type") String type, Principal principal){
-
-        User adminFan  = userService.findByUsername(principal.getName());
-        UsedProp usedProp = null;
+    public ResponseEntity<UsedProp> approveDecline(@PathVariable Long id, @RequestParam("type") String type, Principal principal) {
+        User adminFan = userService.findByUsername(principal.getName());
+        UsedProp usedProp;
         try {
             usedProp = usedPropService.approveDecline(id, type, adminFan);
         } catch (Exception e) {
             return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
+        if (usedProp == null) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         return new ResponseEntity<>(usedProp, HttpStatus.OK);
 
     }
+
     @RequestMapping(
             value = "/{usedPropId}/accept-bid/{acceptedBidId}",
             method = RequestMethod.GET)
-    public ResponseEntity<Bid> acceptBid(@PathVariable Long usedPropId, @PathVariable Long acceptedBidId){
-       boolean success = usedPropService.acceptBid(usedPropId,acceptedBidId);
-       if(success){
-           return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-       }
+    public ResponseEntity<Bid> acceptBid(Principal principal, @PathVariable Long usedPropId, @PathVariable Long acceptedBidId) {
+        boolean success = usedPropService.acceptBid(principal.getName(), usedPropId, acceptedBidId);
+        if (success) {
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
         return new ResponseEntity<>(HttpStatus.FORBIDDEN);
 
     }
+
     @RequestMapping(
             method = RequestMethod.POST,
             consumes = MediaType.APPLICATION_JSON_VALUE,
             produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<?> createPropUsed(Principal principal, @Valid @RequestBody UsedProp usedProp, Errors errors) {
-        if(errors.hasErrors()){
+        if (errors.hasErrors()) {
             return new ResponseEntity<String>(errors.getAllErrors().toString(), HttpStatus.BAD_REQUEST);
         }
-        usedProp = usedPropService.createUsedProp(principal.getName(),usedProp);
+        usedProp = usedPropService.createUsedProp(principal.getName(), usedProp);
         return new ResponseEntity<>(usedProp, HttpStatus.CREATED);
     }
 
@@ -132,8 +144,17 @@ public class UsedPropController {
     @RequestMapping(
             value = "/{id}",
             method = RequestMethod.DELETE)
-    public ResponseEntity<UsedProp> deletePropUsed(@PathVariable("id") Long id) {
-        usedPropService.delete(id);
+    public ResponseEntity<UsedProp> deletePropUsed(Principal principal, @PathVariable("id") Long id) {
+        User user = userService.findByUsername(principal.getName());
+        UsedProp usedProp = usedPropService.findOne(id);
+        if ((usedProp.getUser() != user) || (usedProp.getUser().getRole() != Role.ADMINFAN)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+        try {
+            usedPropService.delete(id);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 }
